requirements.txt
python-telegram-bot==21.3
anthropic==0.28.0
apify-client==1.7.1
apscheduler==3.10.4
python-dotenv==1.0.1
